[core][pruning][be] Rename sparsifier folder to pruner
Summary:
att

Test Plan:
```
python test/test_ao_sparsity.py
```

[ghstack-poisoned]
