+++ torch/ao/pruning/__init__.py
@@ -2,10 +2,10 @@
 from ._mappings import get_dynamic_sparse_quantized_mapping
 from ._mappings import get_static_sparse_quantized_mapping
 
-# Sparsifier
-from .sparsifier.base_pruner import BasePruner
-from .sparsifier.weight_norm_pruner import WeightNormPruner
-from .sparsifier.nearly_diagonal_pruner import NearlyDiagonalPruner
+# Pruner
+from .pruner.base_pruner import BasePruner
+from .pruner.weight_norm_pruner import WeightNormPruner
+from .pruner.nearly_diagonal_pruner import NearlyDiagonalPruner
 
 # Scheduler
 from .scheduler.base_scheduler import BaseScheduler
@@ -13,7 +13,7 @@ from .scheduler.lambda_scheduler import LambdaSL
 from .scheduler.cubic_scheduler import CubicSL
 
 # Parametrizations
-from .sparsifier.utils import FakeSparsity
-from .sparsifier.utils import module_to_fqn
-from .sparsifier.utils import fqn_to_module
-from .sparsifier.utils import get_arg_info_from_tensor_fqn
+from .pruner.utils import FakeSparsity
+from .pruner.utils import module_to_fqn
+from .pruner.utils import fqn_to_module
+from .pruner.utils import get_arg_info_from_tensor_fqn
